JENKINS-12237 : Failed deleting previous generated IPA results in a failed build Don't try to delete if the the directory doesn't exist.

+++ src/main/java/au/com/rayh/XCodeBuilder.java
@@ -386,9 +386,13 @@ public class XCodeBuilder extends Builder {
         // Package IPA
         if (buildIpa) {
 
-            listener.getLogger().println(Messages.XCodeBuilder_cleaningIPA());
-            for (FilePath path : buildDirectory.list("*.ipa")) {
-                path.delete();
+            if (buildDirectory.exists()) {
+                listener.getLogger().println(Messages.XCodeBuilder_cleaningIPA());
+                for (FilePath path : buildDirectory.list("*.ipa")) {
+                    path.delete();
+                }
+            } else {
+                listener.getLogger().println(Messages.XCodeBuilder_NotExistingDirToCleanIPA(buildDirectory.absolutize().getRemote()));
             }
 
             listener.getLogger().println(Messages.XCodeBuilder_packagingIPA());
